feat: build 오류 수정 및 complete 페이지 서버 컴포넌트로 분리
